LayoutTests/wasm/iframe-parent-postmessage-expected.txt

Test that expected Wasm Module can be sent over window.parent.postMessage from iframe.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS () => value is 35010
